refactored file structure of application
