+++ .gitignore
@@ -2,3 +2,4 @@ basketball-v0-906cb-firebase-adminsdk-s3za7-1d68175be0.json
 venv/
 notes.txt
 .env
+__pycache__/
